feat: Adicionando autenticação jwt na api e adicionando base validations.

+++ src/MeuMenu.Domain/Services/ProdutoService.cs
@@ -1,20 +1,29 @@
 ﻿using MeuMenu.Domain.Interfaces.Repository;
 using MeuMenu.Domain.Interfaces.Services;
 using MeuMenu.Domain.Models;
+using MeuMenu.Domain.Services.Base;
 
 namespace MeuMenu.Domain.Services;
 
 public class ProdutoService : IProdutoService
 {
     private readonly IProdutoRepository  _produtoRepository;
+    private readonly NegocioService _negocioService;
 
-    public ProdutoService(IProdutoRepository produtoRepository)
+    public ProdutoService(
+        IProdutoRepository produtoRepository, 
+        NegocioService negocioService)
     {
         _produtoRepository = produtoRepository;
+        _negocioService = negocioService;
     }
 
     public async Task<Produto> AdicionarProdutoAsync(Produto produto)
     {
+        // TODO: Criar validações e adicionar ao produto
+        await _negocioService.ExecutarValidacao();
+        if (!_negocioService.EhValido()) return produto;
+        
         await _produtoRepository.Adicionar(produto);
 
         return produto;
@@ -22,6 +31,10 @@ public class ProdutoService : IProdutoService
 
     public async Task<Produto> AtualizarProdutoAsync(Produto produto)
     {
+        // TODO: Criar validações e adicionar ao produto
+        await _negocioService.ExecutarValidacao();
+        if (!_negocioService.EhValido()) return produto;
+        
         await _produtoRepository.Atualizar(produto);
 
         return produto;
@@ -29,6 +42,9 @@ public class ProdutoService : IProdutoService
 
     public async Task ExcluirProdutoAsync(Guid produtoId)
     {
+        // TODO: Criar validações e adicionar ao produto
+        await _negocioService.ExecutarValidacao();
+        if (!_negocioService.EhValido()) return;
         await _produtoRepository.Remover(produtoId);
     }
